+++ poc/12_advection/marching_cubes/lib.h
@@ -1,11 +1,11 @@
 struct Vec {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 };
 struct Vec;
-float GetOffset(float, float, float);
+double GetOffset(double, double, double);
 void GetColor(struct Vec*, struct Vec*);
 void PrintHelp(void);
-void GetNormal(struct Vec*, float, float, float, double(*f)(double, double, double, void*), void*);
+void GetNormal(struct Vec*, double, double, double, double(*f)(double, double, double, void*), void*);
